Fixed layouts on different pages, fixed bugs and added success/error message

+++ Tweeter/app/Http/Controllers/TweetController.php
@@ -21,19 +21,27 @@ class TweetController extends Controller
             $tweet->content = $request->content;
             $tweet->save();
 
-            return Redirect::back();
+            return Redirect::back()->with('messageSuccess', 'Tweet successfully created!');
         }else {
-            return redirect('/login');
+            return redirect('/login')->with('messageError', 'Please Login to continue!');
         }
     }
 
     function deleteTweet(Request $request){
         if(Auth::check()){
-            \App\Tweet::destroy($request->tweetId);
+            $tweet = \App\Tweet::find($request->tweetId);
 
-            return redirect('/home');
+            if($tweet->user_id == Auth::user()->id){
+                \App\Tweet::destroy($request->tweetId);
+                $message = 'Tweet successfully deleted!';
+                $messageType = 'messageSuccess';
+            } else {
+                $message = 'Tweet belongs to a different user. Unable to delete!';
+                $messageType = 'messageError';
+            }
+            return redirect('/home')->with($messageType, $message);
         }else{
-            return redirect('/login');
+            return redirect('/login')->with('messageError', 'Please Login to continue!');
         }
     }
 
@@ -49,13 +57,24 @@ class TweetController extends Controller
 
             return redirect('/tweet/view/'.$request->tweetId);
         }else{
-            return redirect('/login');
+            return redirect('/login')->with('messageError', 'Please Login to continue!');
         }
     }
 
     function editTweetView(Request $request){
-        $tweet =\App\Tweet::orderBy('created_at', 'asc')->find($request->tweetId);
-        return view('tweeter_tweet_edit')->with(['tweet'=> $tweet]);
+        if(Auth::check()){
+            $tweet =\App\Tweet::orderBy('created_at', 'asc')->find($request->tweetId);
+            if($tweet == null){
+               return redirect('/home')->with('messageError', 'Tweet does not exist!');
+            }
+            if($tweet->user_id != Auth::user()->id){
+                return redirect('/home')->with('messageError', 'Tweet does not belong to you to edit!');
+            }
+
+            return view('tweeter_tweet_edit')->with(['tweet'=> $tweet]);
+        } else {
+            return redirect('/login')->with('messageError', 'Please Login to continue!');
+        }
     }
 
     function likeTweet(Request $request){
@@ -64,9 +83,9 @@ class TweetController extends Controller
             $like->user_id = Auth::user()->id;
             $like->tweet_id = $request->tweetId;
             $like->save();
-            return Redirect::back();
+            return Redirect::back()->with('messageSuccess', 'Liked tweet successfully.');
         }else {
-            return redirect('/login');
+            return redirect('/login')->with('messageError', 'Please Login to continue!');
         }
     }
 
@@ -75,15 +94,24 @@ class TweetController extends Controller
             \App\Like::where('user_id',Auth::user()->id)
                         ->where('tweet_id', $request->tweetId)
                         ->delete();
-            return Redirect::back();
+            return Redirect::back()->with('messageSuccess', 'Unliked tweet successfully.');
         }else {
-            return redirect('/login');
+            return redirect('/login')->with('messageError', 'Please Login to continue!');
         }
     }
 
     function viewProfileTweet($id){
-        $tweet =\App\Tweet::orderBy('created_at', 'asc')->find($id);
-        return view('tweeter_tweet_view', ['tweet'=> $tweet]);
+        if(Auth::check()){
+            $tweet =\App\Tweet::orderBy('created_at', 'asc')->find($id);
+            if($tweet === null){
+                return redirect('home');
+            } else {
+                return view('tweeter_tweet_view', ['tweet'=> $tweet]);
+            }
+        } else {
+            return redirect('/login')->with('messageError', 'Please Login to continue!');
+        }
+
     }
 
     function commentOnTweet(Request $request){
@@ -98,19 +126,28 @@ class TweetController extends Controller
             $comment->content = $request->comment;
             $comment->save();
 
-            return Redirect::back();
+            return Redirect::back()->with('messageSuccess', 'Comment added successfully!');
         } else {
-            return redirect('/login');
+            return redirect('/login')->with('messageError', 'Please Login to continue!');
         }
     }
 
     function commentDelete(Request $request){
         if(Auth::check()){
-            \App\Comment::destroy($request->commentId);
+            $comment = \App\Comment::find($request->commentId);
 
-            return Redirect::back();
+            if($comment->user_id == Auth::user()->id){
+                \App\Comment::destroy($request->commentId);
+                $message = 'Comment deleted successfully';
+                $messageType = 'messageSuccess';
+            } else {
+                $message = 'Comment does not belong to you to delete!';
+                $messageType = 'messageError';
+            }
+
+            return Redirect::back()->with($messageType, $message);
         } else {
-            return redirect('/login');
+            return redirect('/login')->with('messageError', 'Please Login to continue!');
         }
     }
 
@@ -124,16 +161,27 @@ class TweetController extends Controller
             $comment->content = $request->content;
             $comment->save();
 
-            return redirect('/tweet/view/'.$request->tweetId);
+            return redirect('/tweet/view/'.$request->tweetId)->with('messageSuccess', 'Comment edited successfully!');
 
         }else{
-            return redirect('login');
+            return redirect('login')->with('messageError', 'Please Login to continue!');
         }
     }
 
     function commentEditView(Request $request){
-        $tweet = \App\Tweet::find($request->tweetId);
+        if(Auth::check()){
+            $tweet = \App\Tweet::find($request->tweetId);
+            if($tweet == null){
+                return redirect('/home')->with('messageError', 'Comment does not exist!');
+            }
+            if($tweet->user_id != Auth::user()->id){
+                return redirect('/home')->with('messageError', 'Comment does not belong to you to edit!');
+            }
+            return view('tweeter_tweet_comment_edit')->with(['tweet' => $tweet, 'commentId' => $request->commentId]);
+        } else {
+            return redirect('/login')->with('messageError', 'Please Login to continue!');
+        }
 
-        return view('tweeter_tweet_comment_edit')->with(['tweet' => $tweet, 'commentId' => $request->commentId]);
+
     }
 }
